requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-core
langchain-openai
openai
pydantic
python-dotenv
tiktoken